Split strain generation portion of StrainSkill into separate class
